agent_service/requirements.txt
langgraph
langchain
fastapi
uvicorn[standard]
openai==1.14
redis
langfuse
python-dotenv